… lib/Doctrine/Annotations/DocParser.php
… tests/Doctrine/Tests/Annotations/Metadata/TransientMetadataCollectionTest.php