@Ocramius 2019-02-26
Also: needs a performance check now that #248 is merged