@Majkl578 2019-02-26
This branch (already rebased):
```
\Doctrine\Performance\Annotations\DocLexerPerformanceBench

    benchMethod.............................I4 [μ Mo]/r: 7.067 7.086 (μs) [μSD μRSD]/r: 0.065μs 0.92%
    benchClass..............................I4 [μ Mo]/r: 10.108 10.017 (μs) [μSD μRSD]/r: 0.189μs 1.87%

\Doctrine\Performance\Annotations\DocParsePerformanceBench

    benchMethodParsing......................I4 [μ Mo]/r: 36.381 34.131 (μs) [μSD μRSD]/r: 4.568μs 12.56%
    benchClassParsing.......................I4 [μ Mo]/r: 32.458 31.741 (μs) [μSD μRSD]/r: 1.462μs 4.50%

\Doctrine\Performance\Annotations\PhpParserPerformanceWithoutShortCutBench

    bench...................................I4 [μ Mo]/r: 82.972 82.738 (μs) [μSD μRSD]/r: 0.507μs 0.61%

\Doctrine\Performance\Annotations\PhpParserPerformanceWithShortCutBench

    bench...................................I4 [μ Mo]/r: 88.515 88.243 (μs) [μSD μRSD]/r: 1.227μs 1.39%

\Doctrine\Performance\Annotations\ReadPerformanceBench

    bench...................................I4 [μ Mo]/r: 32.372 32.235 (μs) [μSD μRSD]/r: 0.314μs 0.97%

7 subjects, 35 iterations, 2,900 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 6.988 [41.410 40.884] 7.162 (μs)
⅀T: 1,449.365μs μSD/r 1.190μs μRSD/r: 3.260%
```

Master:
```
\Doctrine\Performance\Annotations\DocLexerPerformanceBench

    benchMethod.............................I4 [μ Mo]/r: 7.103 7.055 (μs) [μSD μRSD]/r: 0.093μs 1.31%
    benchClass..............................I4 [μ Mo]/r: 10.043 10.020 (μs) [μSD μRSD]/r: 0.044μs 0.44%

\Doctrine\Performance\Annotations\DocParsePerformanceBench

    benchMethodParsing......................I4 [μ Mo]/r: 23.252 23.183 (μs) [μSD μRSD]/r: 0.367μs 1.58%
    benchClassParsing.......................I4 [μ Mo]/r: 23.165 23.099 (μs) [μSD μRSD]/r: 0.148μs 0.64%

\Doctrine\Performance\Annotations\PhpParserPerformanceWithoutShortCutBench

    bench...................................I4 [μ Mo]/r: 83.942 83.336 (μs) [μSD μRSD]/r: 1.254μs 1.49%

\Doctrine\Performance\Annotations\PhpParserPerformanceWithShortCutBench

    bench...................................I4 [μ Mo]/r: 89.248 89.752 (μs) [μSD μRSD]/r: 1.226μs 1.37%

\Doctrine\Performance\Annotations\ReadPerformanceBench

    bench...................................I4 [μ Mo]/r: 23.663 23.359 (μs) [μSD μRSD]/r: 0.622μs 2.63%

7 subjects, 35 iterations, 2,900 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 7.030 [37.202 37.115] 7.276 (μs)
⅀T: 1,302.083μs μSD/r 0.536μs μRSD/r: 1.352%
```