@Majkl578 2019-02-26
> Overall looking good, what's the likeliness of this introducing a BC break, and should it target 1.x?

I wouldn't push it to 1.x at this state. It doesn't change public API (I reverted removal of `Target::$literal`) however it does remove the _internal_ `static` annotations metadata registry - this makes i.e. ORM test suite slower (4.8s -> 5.3s). We could keep it `static` in 1.x though.