@Ocramius 2019-02-26
The bench for `benchMethodParsing` is quite worrisome though: 36% slower there.

Is there any hot path that we can inspect, or adding utility API on the builder to reduce hops?