@alcaeus 2019-03-05
As per https://github.com/doctrine/annotations/pull/253#issuecomment-469672292, we've decided to not apply doctrine/coding-standard for now. @Majkl578 I've marked style comments in my review above as resolved, feel free to ignore that feedback unless it bothers you personally 😉