@Majkl578 2019-03-05
Addressed review comments from @alcaeus. Also did two more changes:
- removed ArrayAccess, it was superfluous and didn't fit very well into the API;
- extracted MetadataCollection into interface, it could serve as one of few extension points for consumers (for possible caching).

Metadata:
```
\Doctrine\Performance\Annotations\DocLexerPerformanceBench

    benchMethod.............................I4 [μ Mo]/r: 7.058 7.076 (μs) [μSD μRSD]/r: 0.044μs 0.62%
    benchClass..............................I4 [μ Mo]/r: 10.018 10.060 (μs) [μSD μRSD]/r: 0.059μs 0.59%

\Doctrine\Performance\Annotations\DocParsePerformanceBench

    benchMethodParsing......................I4 [μ Mo]/r: 29.042 29.096 (μs) [μSD μRSD]/r: 0.119μs 0.41%
    benchClassParsing.......................I4 [μ Mo]/r: 28.667 28.728 (μs) [μSD μRSD]/r: 0.131μs 0.46%

\Doctrine\Performance\Annotations\PhpParserPerformanceWithoutShortCutBench

    bench...................................I4 [μ Mo]/r: 83.974 84.338 (μs) [μSD μRSD]/r: 0.527μs 0.63%

\Doctrine\Performance\Annotations\PhpParserPerformanceWithShortCutBench

    bench...................................I4 [μ Mo]/r: 88.739 88.268 (μs) [μSD μRSD]/r: 0.828μs 0.93%

\Doctrine\Performance\Annotations\ReadPerformanceBench

    bench...................................I4 [μ Mo]/r: 27.972 27.578 (μs) [μSD μRSD]/r: 0.822μs 2.94%

7 subjects, 35 iterations, 2,900 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 6.988 [39.353 39.306] 7.108 (μs)
⅀T: 1,377.349μs μSD/r 0.361μs μRSD/r: 0.940%
```

Master:
```
\Doctrine\Performance\Annotations\DocLexerPerformanceBench

    benchMethod.............................I4 [μ Mo]/r: 7.086 7.047 (μs) [μSD μRSD]/r: 0.063μs 0.89%
    benchClass..............................I4 [μ Mo]/r: 10.058 10.006 (μs) [μSD μRSD]/r: 0.106μs 1.05%

\Doctrine\Performance\Annotations\DocParsePerformanceBench

    benchMethodParsing......................I4 [μ Mo]/r: 22.700 22.633 (μs) [μSD μRSD]/r: 0.211μs 0.93%
    benchClassParsing.......................I4 [μ Mo]/r: 23.049 22.871 (μs) [μSD μRSD]/r: 0.368μs 1.59%

\Doctrine\Performance\Annotations\PhpParserPerformanceWithoutShortCutBench

    bench...................................I4 [μ Mo]/r: 84.378 84.647 (μs) [μSD μRSD]/r: 0.559μs 0.66%

\Doctrine\Performance\Annotations\PhpParserPerformanceWithShortCutBench

    bench...................................I4 [μ Mo]/r: 88.188 87.933 (μs) [μSD μRSD]/r: 0.656μs 0.74%

\Doctrine\Performance\Annotations\ReadPerformanceBench

    bench...................................I4 [μ Mo]/r: 23.430 23.196 (μs) [μSD μRSD]/r: 0.472μs 2.02%

7 subjects, 35 iterations, 2,900 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 7.016 [36.984 36.905] 7.180 (μs)
⅀T: 1,294.449μs μSD/r 0.348μs μRSD/r: 1.126%
```