@Majkl578 2019-03-05
100% coverage in Metadata namespace. 😃