@Majkl578 2019-03-06
> Do you want to allow constructing exceptions without usage of the factories or should the constructor be private?

Right now we allow it in Migrations and ORM master. Probably not a big deal, note that proper exception structure is yet to be introduced (converting AnnotationException to interface etc.).

> Since there are a lot of final classes, this becomes very dependent on our own classes without the option to properly customise anything.

This is intentional. Most of the parts of Annotations should not be extended and those that could should be carefully considered.
I'd like to hear opinion from @Ocramius here, in my opinion the more we keep closed to midifications, the better, especially for such a widely used library.
Of course we could reconsider this later, before 2.0 or even after - unfinalizing a class should be BC-okay.